% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{bojack_palette}
\alias{bojack_palette}
\title{Bojack character color palette generator}
\usage{
bojack_palette(name, n)
}
\arguments{
\item{name}{Name of desired character or palette. Choices are
\code{Bojack}, \code{Todd}, \code{Diane},
\code{MrPeanutButter}, \code{PrincessCarolyn},
\code{Hollyhock}, \code{SarahLynn}, \code{HorsinAroundBojack},
\code{ViewFromHalfwayDown}
@importFrom graphics rgb rect par image text}
}
\value{
A vector of hex colors
}
\description{
These are palettes for character from the TV series Bojack Horseman
}
\keyword{colors}
