% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{bojack_palettes}
\alias{bojack_palettes}
\title{Complete list of available character palettes}
\format{
An object of class \code{list} of length 9.
}
\usage{
bojack_palettes
}
\description{
Use \code{\link{bojack_palette}} to construct palettes.
}
\keyword{datasets}
